# contracts/openapi/catalog-service.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "catalog-service",
    "description" : "Produto, opções e disponibilidade qualitativa. Toda rota exige token emitido pelo identity-service, e a autorização é resolvida no merchant-service com o token de quem pediu, encaminhado (ADR-045).",
    "version" : "v1"
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "paths" : {
    "/api/v1/merchants/{estabelecimentoId}/catalog/produtos/{produtoId}/grupos/{grupoId}/opcoes/{opcaoId}/disponibilidade" : {
      "put" : {
        "tags" : [ "produto-controller" ],
        "summary" : "Marca a disponibilidade de uma opção do produto",
        "description" : "Mesmo carimbo e mesmas recusas da marcação do produto. Devolve o produto recalculado, porque marcar uma opção pode derrubar o vendavel.",
        "operationId" : "marcarOpcao",
        "parameters" : [ {
          "name" : "estabelecimentoId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "produtoId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "grupoId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "opcaoId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MarcacaoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "O produto recalculado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProdutoResumoResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Corpo inválido, ou opção que não pertence a este produto"
          },
          "401" : {
            "description" : "Sem token, ou token inválido"
          },
          "403" : {
            "description" : "Sem vínculo, sem ALTERAR_PRODUTO, produto de outra loja ou merchant indisponível — a mesma recusa"
          },
          "409" : {
            "description" : "ESGOTADO_HOJE numa loja que não abre por horário (ADR-049), ou o produto mudou durante a marcação (ADR-052)"
          },
          "503" : {
            "description" : "O merchant não respondeu sobre o expediente"
          }
        }
      }
    },
    "/api/v1/merchants/{estabelecimentoId}/catalog/produtos/{produtoId}/disponibilidade" : {
      "put" : {
        "tags" : [ "produto-controller" ],
        "summary" : "Marca a disponibilidade do produto",
        "description" : "O carimbo é do servidor: o instante pelo relógio do serviço e o expediente perguntado ao merchant (ADR-049). Exige ALTERAR_PRODUTO. 409 quando a loja não abre por horário e o estado é ESGOTADO_HOJE. Devolve o produto recalculado.",
        "operationId" : "marcarProduto",
        "parameters" : [ {
          "name" : "estabelecimentoId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "produtoId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MarcacaoRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "O produto recalculado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProdutoResumoResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Corpo inválido, ou estado que o modo do produto não tem (SEM_CONTROLE não acaba)"
          },
          "401" : {
            "description" : "Sem token, ou token inválido"
          },
          "403" : {
            "description" : "Sem vínculo, sem ALTERAR_PRODUTO, produto de outra loja ou merchant indisponível — a mesma recusa"
          },
          "409" : {
            "description" : "ESGOTADO_HOJE numa loja que não abre por horário (ADR-049), ou o produto mudou durante a marcação (ADR-052)"
          },
          "503" : {
            "description" : "O merchant não respondeu sobre o expediente"
          }
        }
      }
    },
    "/api/v1/merchants/{estabelecimentoId}/catalog/produtos" : {
      "get" : {
        "tags" : [ "produto-controller" ],
        "summary" : "Lista os produtos publicados do estabelecimento",
        "description" : "Exige vínculo ativo com VER_PRODUTO, resolvido no merchant-service com o token de quem pediu. Sem vínculo, sem permissão, loja inexistente e merchant indisponível devolvem a mesma recusa.",
        "operationId" : "listar",
        "parameters" : [ {
          "name" : "estabelecimentoId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Zero-based page index (0..N)",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 0,
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "The size of the page to be returned",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "default" : 20,
            "minimum" : 1
          }
        }, {
          "name" : "sort",
          "in" : "query",
          "description" : "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
          "required" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A página de produtos publicados",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PaginaResponseProdutoResumoResponse"
                }
              }
            }
          },
          "400" : {
            "description" : "Identificador ou paginação malformados"
          },
          "401" : {
            "description" : "Sem token, ou token inválido"
          },
          "403" : {
            "description" : "Sem vínculo, sem VER_PRODUTO, loja inexistente ou merchant indisponível — a mesma recusa"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "MarcacaoRequest" : {
        "type" : "object",
        "description" : "O novo estado de disponibilidade. O carimbo é do servidor.",
        "properties" : {
          "estado" : {
            "type" : "string",
            "description" : "DISPONIVEL, ACABANDO, ESGOTADO_HOJE ou ESGOTADO_INDETERMINADO.",
            "enum" : [ "DISPONIVEL", "ACABANDO", "ESGOTADO_HOJE", "ESGOTADO_INDETERMINADO" ],
            "example" : "ESGOTADO_HOJE"
          }
        },
        "required" : [ "estado" ]
      },
      "ProdutoResumoResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "categoriaId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "nome" : {
            "type" : "string"
          },
          "precoBase" : {
            "type" : "number"
          },
          "ordem" : {
            "type" : "integer",
            "format" : "int32"
          },
          "disponibilidade" : {
            "type" : "string",
            "enum" : [ "DISPONIVEL", "ACABANDO", "ESGOTADO_HOJE", "ESGOTADO_INDETERMINADO" ]
          },
          "vendavel" : {
            "type" : "boolean"
          }
        }
      },
      "PaginaResponseProdutoResumoResponse" : {
        "type" : "object",
        "properties" : {
          "conteudo" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProdutoResumoResponse"
            }
          },
          "pagina" : {
            "type" : "integer",
            "format" : "int32"
          },
          "tamanho" : {
            "type" : "integer",
            "format" : "int32"
          },
          "total" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "description" : "Token do identity-service. Seis claims, sem papel nem permissão — a permissão vem do merchant, não do token.",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}